requirements.txt
django
bs4
requests
python-dateutil
networkx
markdownify
google-genai